requirements.txt
beautifulsoup4>=4.9.3
sheetfu>=1.6.1
termcolor>=1.1.0
requests>=2.25.1
google-cloud-storage>=2.3.0
